requirements.txt
discord.py>=2.4.0
flask>=3.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
mypy>=1.0.0
